Fix two bugs in TwoBound merchant
Fixed a bug that made the merchant crash when the producer offered new products and the product was sold on the market before TwoBound received it itself.
fixed a bug that lead to a crash due to letting the merchant sleep for a negativ number of seconds.
Further more, commented out some annoying debug prints.